Create README.md
This study proposes an innovative method to optimize the inspection paths of offshore wind farms, integrating the K-means clustering algorithm and Genetic Algorithm (GA). The paper formulates the inspection path planning problem as a Multiple Traveling Salesman Problem (mTSP) and utilizes the advantages of the K-means clustering algorithm in distance similarity to effectively group the positions of wind turbines, optimizing the inspection schedules of vessels. Subsequently, leveraging the powerful optimization capability and robustness of the Genetic Algorithm, the study further refines the search for the optimal inspection routes, aiming to achieve cost reduction.